klifurmot-backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.40.3
botocore==1.40.3
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
coverage==7.10.7
cryptography==46.0.3
daphne==4.2.0
dj-database-url==3.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-expiring-token==1.0.3
django-extensions==4.1
django-ses==4.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
google-auth==2.40.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
MarkupSafe==3.0.3
msgpack==1.1.1
oauthlib==3.2.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
sendgrid==6.12.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
zope.interface==7.2
